add charch

+++ chain/temple.go
@@ -2,6 +2,6 @@ package chain
 
 type ChainOp interface {
 	DoTransaction(string, string, string) (string, error)
-	QueryTransaction(string) (int64, error)
+	QueryTransaction(string) (int, error)
 	GenKeyPair() (string, string, error)
 }
